+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/structure_fields/AddFieldDropdown.tsx
@@ -15,7 +15,7 @@ import {
 	FIELD_TYPES,
 	FIELD_TYPE_LABEL,
 	getDefaultField,
-} from '../../../structure_builder/utils/fieldType';
+} from '../../../structure_builder/utils/field';
 
 export default function AddFieldDropdown({
 	triggerType = 'text',
