+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/types/ObjectDefinition.d.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {FieldBusinessType} from '../utils/fieldType';
+import {FieldBusinessType} from '../utils/field';
 
 export type ObjectField = {
 	businessType: FieldBusinessType;
